Atmosphere and Pedal Hi-Hat tweaks
Supplice work determined the Atmosphere patch sounded goofy (and was a
Paul Radek holdover) so it was destroyed and recreated with the SC-55
patch in mind. Good riddance to that old patch, I hated it with a
passion.

Pedal Hat has also been redone and no longer sounds like  "vwee fwee
fee" so that's good.