Adding even more Instruments
Okay so I found a "bug" in libADLMIDI in the process so I'll document that here - Some SC-88 GS midis use "BankSel" for selecting banks, not LSB. ADLMidi actually takes the MSB, multiplies that by 256 (?????) and then adds the LSB number onto that. if BANKSEL is active, LSB needs to be ignored completely, because it's garbage at that point. I think there was some extra GS silliness in that, like variants of voices on bank 0, but that opens up a *huge* can of worms!!

Which is why GM2 standard exists - nonsense like this!