"Old" toms reintroduced with tweaks
I didn't want to do this, but given how tom-heavy some of the Square Ep 2 midis are, it sounded like off-key notes. I know the older ones were "more correct", but in some of the drum fills, it sounded so off. I'll have to figure out what to do there.

The older ones are also way louder.